refactor: 💡 BaseTimeEntity 통합으로 인한 코드 수정
