--- backend/joalarm/src/main/java/com/cupid/joalarm/baseEntity/Auditable.java
@@ -1,12 +1,0 @@
-package com.cupid.joalarm.baseEntity;
-
-import java.time.LocalDateTime;
-
-public interface Auditable {
-    LocalDateTime getCreatedAt();
-    LocalDateTime getUpdatedAt();
-
-    void setCreatedAt(LocalDateTime createdAt);
-
-    void setUpdatedAt(LocalDateTime updatedAt);
-}
