--- backend/joalarm/src/main/java/com/cupid/joalarm/baseEntity/BaseTimeEntity.java
@@ -1,27 +1,0 @@
-package com.cupid.joalarm.baseEntity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Data
-@MappedSuperclass
-@EntityListeners(value = AuditingEntityListener.class)
-public class BaseTimeEntity implements Auditable{
-    @CreatedDate
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-}
