+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/Feed.java
@@ -1,16 +1,32 @@
 package com.cupid.joalarm.feed;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.cupid.joalarm.baseEntity.BaseTimeEntity;
+import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.base.entity.BaseTimeEntity;
 import com.cupid.joalarm.feed.comment.Comment;
 import com.cupid.joalarm.feed.like.Like;
 import com.cupid.joalarm.feed.tag.Tag;
-import com.cupid.joalarm.account.entity.Account;
-import lombok.*;
-
-import javax.persistence.*;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.ArrayList;
 import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Entity
 @Table(name = "feed")
@@ -48,12 +64,12 @@ public class Feed extends BaseTimeEntity {
     private Account account;
 
     @Builder.Default
-    @OneToMany(mappedBy = "feed",cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "feed", cascade = CascadeType.ALL)
     @JsonIgnoreProperties({"feed"})
     private List<Comment> comments = new ArrayList<>();
 
     @Builder.Default
-    @OneToMany(mappedBy = "feed",cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "feed", cascade = CascadeType.ALL)
     @JsonIgnoreProperties({"feed"})
     private List<Like> likes = new ArrayList<>();
