+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/childcomment/ChildComment.java
@@ -1,11 +1,23 @@
 package com.cupid.joalarm.feed.childcomment;
 
 import com.cupid.joalarm.account.entity.Account;
-import com.cupid.joalarm.baseEntity.BaseTimeEntity;
+import com.cupid.joalarm.base.entity.BaseTimeEntity;
 import com.cupid.joalarm.feed.comment.Comment;
-import lombok.*;
-
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Entity
 @Table(name = "childcomment")
