+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/comment/Comment.java
@@ -1,15 +1,29 @@
 package com.cupid.joalarm.feed.comment;
 
-import com.cupid.joalarm.baseEntity.BaseTimeEntity;
-import com.cupid.joalarm.feed.Feed;
 import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.base.entity.BaseTimeEntity;
+import com.cupid.joalarm.feed.Feed;
 import com.cupid.joalarm.feed.childcomment.ChildComment;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Entity
 @Table(name = "comment")
@@ -38,7 +52,7 @@ public class Comment extends BaseTimeEntity {
     private Feed feed;
 
     @Builder.Default
-    @OneToMany(mappedBy = "comment",cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL)
     @JsonIgnoreProperties({"comment"})
     private List<ChildComment> comments = new ArrayList<>();
 }
